requirements.txt
langchain==0.1.16
langchain-community
langchain-huggingface
langchain-groq
chromadb
streamlit
python-dotenv
